% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspaces.R
\name{.get_projects_path}
\alias{.get_projects_path}
\title{Get the current RRLab projects base path}
\usage{
.get_projects_path()
}
\value{
The absolute path to the projects folder
}
\description{
Get the current RRLab projects base path
}
