% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspaces.R
\name{.load_project_structure}
\alias{.load_project_structure}
\title{Set up a project folder structure inside the RRLab projects path}
\usage{
.load_project_structure(project_name, branch = "main")
}
\description{
Set up a project folder structure inside the RRLab projects path
}
