% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspaces.R
\name{.setup_project_structure}
\alias{.setup_project_structure}
\title{Set up a project folder structure inside the RRLab projects path}
\usage{
.setup_project_structure(project_name)
}
\arguments{
\item{project_name}{Name of the new project folder}
}
\value{
The full path to the created project
}
\description{
Set up a project folder structure inside the RRLab projects path
}
