% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Idiogram.R
\name{Idiogram}
\alias{Idiogram}
\title{Idiogram}
\usage{
Make Idiogram from dataframe object based on CpGs. Indicates the chromosomal location of which CpGs were found, and its significant level. Only chromosomes 1 untill 22 are supported for now. Can be developed if needed.
}
\arguments{
\item{data_points}{Data.frame must contain: "seqnames", "start", "end" and "pvalue", others are optional. See example.}

\item{s_barwidth}{Plotting (bar). Chromosome bar width}

\item{s_size_line}{Plotting (lines) CpG line size}

\item{s_alpha_lines}{Plotting (lines) CpG line Alpha}

\item{s_offset}{Plotting (points) CpG significance point offset to line}

\item{s_size_point}{Plotting (points) CpG significance point size}
}
\description{
Idiogram
}
\examples{
# This is an example using a randomized set of sites of interest on the genome, then adds random unified sampled P-values. Then shows the plot
npoints = 1000
data_points <- sort(regioneR::createRandomRegions(nregions=npoints, mask=NA))
data_points$pvalue = runif(npoints)
Idiogram(data_points)

}
