% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePCACorrelates.R
\name{MakePCACorrelates}
\alias{MakePCACorrelates}
\title{MakePCACorrelates}
\usage{
MakePCACorrelates(
  Beta,
  pheno,
  output_dir = "./plots/",
  savename = "01",
  var_threshold = 0.9,
  p_threshold = 0.05
)
}
\description{
MakePCACorrelates
}
