% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCC.R
\name{MCC}
\alias{MCC}
\alias{MCC.default}
\alias{MCC.roc}
\title{MCC}
\usage{
MCC(data, ...)

\method{MCC}{default}(data, lev = NULL, model = NULL, showCM = FALSE, Verbose = FALSE)

\method{MCC}{roc}(data, lev = NULL, model = NULL, showCM = FALSE, Verbose = FALSE)
}
\arguments{
\item{data}{Input the dataset object (column 1 = OBS; column 2 = PRED). Additonal 2 columns predictions for class 1 then class 2; see example}

\item{lev}{NULL}

\item{model}{NULL}

\item{showCM}{Show output of confusion matrix & metrics as cat.}
}
\value{
Several metrics to indicate machine learning performance ("MCC","Accuracy","Recall","Precision","F1","CombinedScore")
}
\description{
This function required a dataset with column 1 as observations and column 2 with predictions. Two-class only!
}
\examples{
MCC(data.frame(obs=c(1,1,1,2,2,2),pred=c(1,1,1,1,2,2)))
MCC(data.frame(obs=c("a","a","a","b","b","b"),pred=c("a","a","a","a","b","b"),a=c(0.9,0.8,0.9,0.2,0.3,0.9),b=c(0.1, 0.2, 0.1, 0.8, 0.7, 0.1)),lev=c("a","b"))
# Example high accurate
MCC(data.frame(obs=c(rep(1,1000),rep(2,1000)),pred=c(rep(1,1000),rep(2,1000))))
# Example Inverse accurate
MCC(data.frame(obs=c(rep(2,1000),rep(1,1000)),pred=c(rep(1,1000),rep(2,1000))))
# Example random 
MCC(data.frame(obs=c(round(runif(200000,0,1))),pred=c(round(runif(200000,0,1)))))
# Example high accurate inbalance
MCC(data.frame(obs=c(rep(1,1000),rep(2,10)),pred=c(rep(1,1000),rep(2,10))))
# Example high accurate EXTREME inbalance
MCC(data.frame(obs=c(rep(1,1000),rep(2,10)),pred=c(rep(1,1010))))
# Example moderate accurate inbalance
MCC(data.frame(obs=c(rep(1,1000),round(runif(10,1,2))),pred=c(rep(1,1000),rep(2,10))))

OR with ROC object
data(aSAH)

Rocobj = roc(aSAH$outcome, aSAH$s100b,
   levels=c("Good", "Poor"))
MCC.roc(Rocobj)
}
