% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierCheck.R
\name{OutlierCheck}
\alias{OutlierCheck}
\title{OutlierCheck}
\usage{
OutlierCheck(data, Type = 1)
}
\arguments{
\item{data}{Input the dataset object (row=feature; col=sample).}

\item{Type}{Type indicates sample outliers or feature outliers to be resulted. Defaults to 1 (SAMPLES), can be set to 2 (FEATURES).}
}
\value{
A PCA plot indicative of outlier samples based a set of tests.
}
\description{
This function required a dataset with rows as features and columns as samples.
}
\examples{
OutlierCheck(t(mtcars),1)
}
