% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRdistribution.R
\name{RRDistribution}
\alias{RRDistribution}
\title{RRDistribution}
\usage{
RRDistribution(Data)
}
\arguments{
\item{Data}{Unlisted vector of numbers.}
}
\value{
A distribution estimate with plots.
}
\description{
This function will get a estimate of the distribution of entered input, and indicates what distribution might fit better.
}
\examples{
RRDistribution(Data= iris[,-5])
}
