% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDataset.R
\name{RRFEDataset}
\alias{RRFEDataset}
\title{RRFEDataset}
\usage{
RRFEDataset(SelectDataSet = 1)
}
\value{
dataframe to test the RRFE function to default.
}
\description{
This function sets dataset for RRFE
}
\examples{
dataset1 = RRFEDataset(1)
dataset2 = RRFEDataset(2)
dataset3 = RRFEDataset(3)
}
