% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDefaults.R
\name{RRFEDefaults}
\alias{RRFEDefaults}
\title{RRFEDefaults}
\usage{
RRFEDefaults()
}
\value{
Variables to set the RRFE function to default.
}
\description{
This function sets defaults for RRFE
}
\examples{
dataset = RRFEDefaults()
}
