% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRFilt.R
\name{RRFilt}
\alias{RRFilt}
\title{RRFilt}
\usage{
RRFilt(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
An indication of 'difference' in the given vector. Higher is better diff.
}
\description{
This function gives an indication how big the difference is in a numeric vector. It calculates `abs(mean(x) - median(x)) / sd(x)`, which is near zero for symmetric distributions and larger when the mean and median diverge.
}
\examples{
apply(iris[,-5],2,RRFilt)
}
