% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRgeneEnrichment.R
\name{RRgeneEnrichment}
\alias{RRgeneEnrichment}
\title{RRgeneEnrichment}
\usage{
RRgeneEnrichment(genes, organism = "org.Hs.eg.db", type = "GO",
  pvalueCutoff = 0.05, array.type = "EPIC", ...)
}
\arguments{
\item{genes}{Character vector of gene or CpG identifiers.}

\item{organism}{Annotation package or KEGG organism code.}

\item{type}{Type of enrichment to perform: \code{"GO"} or \code{"KEGG"}.}

\item{pvalueCutoff}{Cutoff for adjusted p-value.}
\item{array.type}{Illumina array type when supplying CpGs.}

\item{...}{Additional arguments passed to the enrichment functions.}
}
\value{
A list with the result table, barplot, dotplot and selection information.
}
\description{
Convenience wrapper around \code{clusterProfiler} to run GO or KEGG
pathway enrichment on a vector of genes.
}
\examples{
\dontrun{
# Basic GO enrichment
genes <- c("BRCA1", "TP53", "EGFR")
RRgeneEnrichment(genes, organism = "org.Hs.eg.db", type = "GO")
## CpG input
RRgeneEnrichment(c("cg00000029", "cg00000108"), type = "GO", array.type = "EPIC")
}
}
