% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspaces.R
\name{set_projects_path}
\alias{set_projects_path}
\title{Set or update the base path for RRLab projects}
\usage{
set_projects_path(path, create = TRUE)
}
\arguments{
\item{path}{The new projects folder path (absolute or relative)}

\item{create}{If TRUE (default), the folder is created if it doesn't exist}
}
\value{
Invisible TRUE
}
\description{
Set or update the base path for RRLab projects
}
