% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsave.R
\name{Rsave}
\alias{Rsave}
\title{Rsave}
\usage{
Rsave(object, prefix = "", postfix = "", file_location = s_saveloc_qc, ...)
}
\arguments{
\item{object}{Object to save or the name of the object.}
\item{prefix}{Optional prefix for the file name.}
\item{postfix}{Optional postfix for the file name.}
\item{file_location}{Directory where the file will be saved. Defaults to \code{s_saveloc_qc}.}
\item{...}{Additional arguments passed to \code{qs::qsavem}.}
}
\value{
Invisibly returns the path used to save the object.
}
\description{
Convenience wrapper around \code{qs::qsavem} that saves an object by its R name. You can specify a prefix, postfix and file location. The default directory is \code{s_saveloc_qc}.
}
\examples{
a <- 1:5
Rsave(a)
Rsave(a, prefix = "08A", postfix = "cohort")
}
