% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspaces.R
\name{Rsave}
\alias{Rsave}
\title{Rsave}
\usage{
Rsave(object_name, prefix = "", postfix = "", file_location = NULL, ...)
}
\arguments{
\item{prefix}{Optional prefix for the file name.}

\item{postfix}{Optional postfix for the file name.}

\item{file_location}{Directory where the file will be saved. If `NULL`,
`s_saveloc_qc` will be used when available.}

\item{...}{Additional arguments passed to `qs::qsavem`.}

\item{object}{Object to save or the name of the object.}
}
\value{
Invisibly returns the file path used.
}
\description{
Convenience wrapper around `qs::qsavem` that saves an object using its
name in the calling environment. A prefix and postfix can be supplied to
modify the file name. If `file_location` is not supplied, `s_saveloc_qc`
will be used when available.
}
\examples{
\dontrun{
a <- 1:5
Rsave(a)
Rsave(a, prefix = "08A", postfix = "cohort")
}
}
